Revert "Fix verify key"
